requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
ax-platform==0.2.5.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.5
botorch==0.6.4
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.4
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
einops==0.2.0
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
fvcore==0.1.5.post20221221
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpytorch==1.9.1
grpcio==1.43.0
huggingface-hub==0.11.1
hydra-core==1.2.0
idna==3.4
imageio==2.24.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
iopath==0.1.10
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linear-operator==0.3.0
lmdb
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
msgpack==1.0.4
multidict==6.0.4
multipledispatch==0.6.0
networkx==3.0
nltk==3.7
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
Pillow==9.0.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
plotly==5.11.0
portalocker==2.6.0
prefetch-generator==1.0.1
protobuf==3.20.1
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
pyDeprecate==0.3.2
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
ray==1.13.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.0.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.15.0
shapely==2.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
streamlit==1.16.0
streamlit-drawable-canvas==0.9.0
# -e git+https://github.com/lynguyenminh/scenetext-api.git@d439c370a8384778222eddd5d729a20eff8080de#egg=strhub&subdirectory=parseq
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
toml==0.10.2
toolz==0.12.0
torch
torchmetrics
torchtext
torchvision
tornado==6.2
tqdm==4.64.1
typeguard==2.13.3
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.39
urllib3==1.26.13
validators==0.20.0
vietocr==0.3.8
virtualenv==20.17.1
watchdog==2.2.1
Werkzeug==2.2.2
yacs==0.1.8
yarl==1.8.2
zipp==3.11.0